requirements.txt
django==5.2.12
channels>=4.3.2
celery>=5.6.2
daphne>=4.2.1
djangorestframework>=3.16.1
djangorestframework-simplejwt>=5.5.1
boto3>=1.34.0
django-storages>=1.14.0
gunicorn>=23.0.0
Pillow>=10.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
razorpay>=1.4.2
redis>=7.3.0
